+++ huaweicloud/provider.go
@@ -1899,8 +1899,6 @@ func Provider() *schema.Provider {
 			"huaweicloud_nat_private_snat_rule":  nat.ResourcePrivateSnatRule(),
 			"huaweicloud_nat_private_transit_ip": nat.ResourcePrivateTransitIp(),
 
-			"huaweicloud_network_acl":              ResourceNetworkACL(),
-			"huaweicloud_network_acl_rule":         ResourceNetworkACLRule(),
 			"huaweicloud_networking_secgroup":      vpc.ResourceNetworkingSecGroup(),
 			"huaweicloud_networking_secgroup_rule": vpc.ResourceNetworkingSecGroupRule(),
 			"huaweicloud_networking_vip":           vpc.ResourceNetworkingVip(),
@@ -2260,6 +2258,8 @@ func Provider() *schema.Provider {
 			"huaweicloud_csbs_backup_policy_v1":          deprecated.ResourceCSBSBackupPolicyV1(),
 			"huaweicloud_csbs_backup_v1":                 deprecated.ResourceCSBSBackupV1(),
 			"huaweicloud_fgs_trigger":                    deprecated.ResourceFunctionGraphTrigger(),
+			"huaweicloud_network_acl":                    deprecated.ResourceNetworkACL(),
+			"huaweicloud_network_acl_rule":               deprecated.ResourceNetworkACLRule(),
 			"huaweicloud_networking_network_v2":          deprecated.ResourceNetworkingNetworkV2(),
 			"huaweicloud_networking_subnet_v2":           deprecated.ResourceNetworkingSubnetV2(),
 			"huaweicloud_networking_floatingip_v2":       deprecated.ResourceNetworkingFloatingIPV2(),
